% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cwr.matrix_out}
\alias{cwr.matrix_out}
\title{This function runs the leave-one-out cross validation}
\usage{
cwr.matrix_out(y = y_sample, x = x_sample, net = net,
  bandwidth = bandwidth, obs = sample_obs, distance.w = distance.w)
}
\description{
This function runs the leave-one-out cross validation
}
