% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{net}
\alias{net}
\title{Twitter Network Data for Donal Trump}
\format{A igraph object with data frame with 5771 vertices and 7229 edges.
The igraph also comes with the following features extracted directly from the Twitter API
\describe{
  \item{name}{name of the user who retweeted Trump}
  \item{ids}{id of the user who retweeted Trump}
  ...
}}
\source{
\url{https://developer.twitter.com/}
}
\usage{
net
}
\description{
A igraph object build after acccessing the Twitter API and collecting the network of
retweets to Donald Trump profile.
}
\keyword{datasets}
