% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pwr.R
\name{pwr.1way}
\alias{pwr.1way}
\title{Estimate the Power of a One-Way ANOVA.}
\usage{
pwr.1way(n, nlevels, mu = 0.3 * 1:nlevels, alpha = 0.05, nsim = 100)
}
\arguments{
\item{n}{The number of observations in each group.}

\item{nlevels}{The number of (levels of the) group(ing variable).}

\item{mu}{The group (-level variable) means.}

\item{alpha}{The level of the test.}

\item{nsim}{The number of simulations on which the power estimate is based.}
}
\value{
A simulation based estimate of the power.
}
\description{
Estimate the Power of a One-Way ANOVA.
}
\examples{
pwr.1way(10, 3, sqrt(0.3)*1:3, nsim=500)
power.anova.test(3, 10, 0.3, 1) # approximately the same
}

