% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pwr.R
\name{pwr.factorial}
\alias{pwr.factorial}
\title{What is the power of a multiway ANOVA?}
\usage{
pwr.factorial(n, design = expand.grid(A = letters[1:3], B = LETTERS[1:3]),
  delta = 0.2, mu = delta * slice.index(array(0, dim = sapply(design,
  nlevels)), 1), alpha = 0.05, nsim = 100)
}
\arguments{
\item{n}{The number(s) of observations of each row of design.}

\item{design}{A data.frame that specifies the design of the multiway factorial structure.}

\item{delta}{The effect size factor for group mean differences specified in mu.}

\item{mu}{An array with dimensions equal to the levels of the columns of design,
specifying the additive effects (this should change to allow for
interaction effects; probably using an lm object).}

\item{alpha}{Level of the tests.}

\item{nsim}{The number of simulations performed for computing the power.}
}
\value{
A vector of simulation based power estimates for each of the (interaction) effects.
}
\description{
What is the power of a multiway ANOVA?
}
\examples{
n = 1:12 * 5
power = sapply(n, pwr.factorial, mu = 0.3 * (row(diag(3))+col(diag(3))))
require(ggplot2)
qplot(n, power[2,], geom=c('point', 'smooth'))
}

