% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwr.R
\name{pwr}
\alias{pwr}
\title{pwr}
\usage{
pwr(net, y, X, bandwidth="cv_pwr")
}
\arguments{
\item{\code{net}:}{an igraph network object.}

\item{\code{y}:}{the dependent variable of the PWR model}

\item{\code{X}:}{a matrix of covariates for the model.}

\item{\code{bandwidht}:}{the bandwidth for the decay function applied to the connected paths. Default uses
the function \code{pwr_cv} to calculate the optimal bandwidth using leave-one-out cross validation}
}
\description{
\code{pwr} estimates the PWR model
}
\examples{
l <- pwr(net=net,y=log(V(net)$in_degree+1),
 X=as.matrix(log(V(net)$out_degree+1)))
}
