% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pwr.R
\name{pwr.ttest}
\alias{pwr.ttest}
\title{Estimate the Power of a T-Test.}
\usage{
pwr.ttest(n, delta, alpha = 0.05)
}
\arguments{
\item{n}{The number of observations in each group.}

\item{delta}{The effect size.}

\item{alpha}{The level of the test.}
}
\value{
The simulation based power estimate.
}
\description{
Estimate the Power of a T-Test.
}
\examples{
pwr.ttest(10,0.3)
}

