% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwr_cv.R
\name{pwr_cv}
\alias{pwr_cv}
\title{pwr_cv}
\usage{
pwr_cv(net, y, X, bandwidth=exp(seq(from=-1, to=3, by=.05)), n=lenght(V(net)))
}
\arguments{
\item{\code{net}:}{an igraph network object.}

\item{\code{y}:}{the dependent variable of the PWR model}

\item{\code{X}:}{a matrix of covariates for the model.}

\item{\code{bandwidht}:}{the values for the bandwidth search. Default searchs over the following values \code{exp(seq(from=-1, to=3, by=.05))}}

\item{\code{n}:}{sample size for the CV search. We suggest the user to reduce the sample to 500 observations for
computational reasons. Default is the entire network.}
}
\value{
it returns a list with two objects: the optimal bandwidth and a graph showing the selection
}
\description{
\code{pwr_cv} performs leave-one-out cross-validation to select the optimal bandwidth to the decay function used
in the PWR estimation.
}
\examples{
l <- cwr.cv_leaveoneout(net=net,y=log(V(net)$in_degree+1),
 X=as.matrix(log(V(net)$out_degree+1)), n=500)
}
